requirements.txt
asgiref==3.2.7
astroid==2.3.3
click==7.1.1
colorama==0.4.3
Django==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
Flask==1.1.1
Flask-MySQLdb==0.2.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
psycopg2==2.8.4
psycopg2-binary==2.8.4
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.11.2
